reorg, update url params to handle new filter
