+++ src/applications/gi-sandbox/constants.js
@@ -47,10 +47,6 @@ export const TABS = Object.freeze({
   location: 'location',
 });
 
-// default state is checked so these will only be present if their corresponding boxes are unchecked
-export const FILTERS_EXCLUDED_FLIP = ['schools', 'employers', 'vettec'];
-export const FILTERS_IGNORE_ALL = ['country', 'state', 'specialMission'];
-
 export const ariaLabels = Object.freeze({
   learnMore: {
     giBillBenefits: 'Learn more about VA education and training programs',
