+++ src/applications/gi-sandbox/containers/search/LocationSearchForm.jsx
@@ -14,7 +14,7 @@ import KeywordSearch from '../../components/search/KeywordSearch';
 import 'mapbox-gl/dist/mapbox-gl.css';
 import Modal from '@department-of-veterans-affairs/component-library/Modal';
 import { useHistory } from 'react-router-dom';
-import { updateUrlParams } from '../../utils/helpers';
+import { updateUrlParams } from '../../selectors/search';
 import { TABS } from '../../constants';
 
 export function LocationSearchForm({
